feat(input): update demo for input
