+++ src/input/_example/index.less
@@ -5,3 +5,18 @@
     }
   }
 }
+
+.t-cell-group + .t-cell-group {
+  margin-top: 16px;
+}
+
+.demo-group {
+  .t-cell-group + .t-cell-group {
+    margin-top: 0;
+  }
+
+  .t-cell-group__title {
+    font-size: 12px;
+    line-height: 20px;
+  }
+}
